Convert portal email store to TS
